@@ -61,18 +64,20 @@
     "posthog-js": "^1.225.0",
     "promise-polyfill": "^8.3.0",
     "ramda": "^0.30.0",
-    "react": "^16.13.1",
-    "react-bootstrap": "^1.0.0-beta.15",
-    "react-content-loader": "^5.1.2",
-    "react-dom": "^16.13.1",
+    "react": "^18.2.14",
+    "react-bootstrap": "^2.0.0",
+    "react-content-loader": "^7.0.0",
+    "react-dom": "^18.2.14",
     "react-dotdotdot": "1.3.1",
-    "react-infinite-scroller": "^1.2.4",
-    "sass": "1.77.8",
+    "react-infinite-scroller": "^1.2.6",
+    "sass": "1.86.0",
     "video.js": "^7.11.8",
     "videojs-transcript-ac": "^0.8.8",
     "videojs-youtube": "^2.6.1"
   },
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@babel/preset-typescript": "^7.25.7",
     "@mitodl/ocw-to-hugo": "1.36.0",
     "@playwright/test": "^1.45.0",
     "@rrweb/types": "2.0.0-alpha.17",
@@ -139,7 +146,7 @@
     "ts-jest": "^27.1.2",
     "ts-loader": "9.3.1",
     "ts-node": "^10.9.1",
-    "typescript": "^4.5.4",
+    "typescript": "^5.4",
     "webpack": "^5.73.0",
     "webpack-bundle-analyzer": "^4.6.1",
     "webpack-cli": "^4.10.0",
