@@ -29,7 +29,7 @@ const config: webpack.Configuration = {
   },
   entry: {
     [entryNames.courseV2]: [
-      fromRoot("./course-v2/assets/course-v2.ts"),
+      fromRoot("./course-v2/assets/course-v2.tsx"),
       fromRoot("./base-theme/assets/index.ts")
     ],
     [entryNames.courseOffline]: [
