@@ -13,7 +13,14 @@ import {
   showSolution
 } from "./js/quiz_multiple_choice"
 import posthog from "posthog-js"
-import { initPostHog } from "../../base-theme/assets/js/posthog"
+import {
+  initPostHog,
+  isFeatureEnabled
+} from "../../base-theme/assets/js/posthog"
+import { QueryClientProvider } from "@tanstack/react-query"
+import { makeQueryClient } from "../../base-theme/assets/js/clients"
+import UserMenu from "../../base-theme/assets/js/components/UserMenu"
+import { createRoot } from "react-dom/client"
 
 export interface OCWWindow extends Window {
   initNanogallery2: () => void
@@ -31,12 +38,24 @@ $(function() {
   checkAnswer()
   showSolution()
   initCourseDrawersClosingViaSwiping()
+  const userMenuContainers = document.querySelectorAll(".user-menu-container")
+  if (userMenuContainers && isFeatureEnabled("ocw-learn-integration")) {
+    for (const userMenuContainer of Array.from(userMenuContainers)) {
+      const queryClient = makeQueryClient()
+      const root = createRoot(userMenuContainer)
+      root.render(
+        <QueryClientProvider client={queryClient}>
+          <UserMenu />
+        </QueryClientProvider>
+      )
+    }
+  }
 })
 
 let nanogallery2Loaded = false
 
 window.initNanogallery2 = () => {
   if (nanogallery2Loaded) return
-  import("./nanogallery2-imports")
+  import("./nanogallery2-imports.js")
   nanogallery2Loaded = true
 }
